Refine webpack split and docs
