+++ package.json
@@ -9,11 +9,14 @@
     "url": "git+https://github.com/ivanlay/dicewarsjs.git"
   },
   "scripts": {
-    "dev": "webpack serve --mode development",
-    "serve": "webpack serve --mode development",
-    "build": "webpack --mode production",
-    "build:analyze": "webpack --mode production --env analyze=true",
-    "analyze": "webpack --mode production --env analyze=true && open dist/bundle-report.html",
+    "dev": "webpack serve --config webpack.modern.js --mode development",
+    "dev:legacy": "webpack serve --config webpack.legacy.js --mode development",
+    "serve": "webpack serve --config webpack.modern.js --mode development",
+    "build": "webpack --config webpack.modern.js --mode production",
+    "build:legacy": "webpack --config webpack.legacy.js --mode production",
+    "build:all": "npm run build && npm run build:legacy",
+    "build:analyze": "webpack --config webpack.modern.js --mode production --env analyze=true",
+    "analyze": "webpack --config webpack.modern.js --mode production --env analyze=true && open dist/bundle-report.html",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
